requirements.txt
asgiref==3.8.1
boto3==1.37.27
botocore==1.37.27
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
idna==3.10
jmespath==1.0.1
python-dateutil==2.9.0.post0
requests==2.32.3
s3transfer==0.11.4
six==1.17.0
sqlparse==0.5.3
urllib3==1.26.20
